refactored small piece of code
