+++ src/pages/PeoplePage.tsx
@@ -1,9 +1,12 @@
-import { useEffect, useState } from "react";
-import { getPeople } from "../api";
-import { NoPeopleMessage, PeopleLoadingError } from "../components/ErrorMessage";
-import { Loader } from "../components/Loader";
-import { PeopleTable } from "../components/Loader/PeopleTable";
-import { Person } from "../types";
+import { useEffect, useState } from 'react';
+import { getPeople } from '../api';
+import {
+  NoPeopleMessage,
+  PeopleLoadingError,
+} from '../components/ErrorMessage';
+import { Loader } from '../components/Loader';
+import { PeopleTable } from '../components/PeopleTable';
+import { Person } from '../types';
 
 export const PeoplePage = () => {
   const [people, setPeople] = useState<Person[] | null>(null);
@@ -16,16 +19,16 @@ export const PeoplePage = () => {
       const data = await getPeople();
 
       setPeople(data);
-    } catch  {
+    } catch {
       setError(true);
     } finally {
       setIsLoading(false);
     }
-  }
+  };
 
   useEffect(() => {
     downloadPeople();
-  }, [])
+  }, []);
 
   return (
     <div className="block">
@@ -34,12 +37,12 @@ export const PeoplePage = () => {
       <div className="box table-container">
         {isLoading && <Loader />}
 
-        {error && <PeopleLoadingError/>}
+        {error && <PeopleLoadingError />}
 
         {people && people.length < 1 && <NoPeopleMessage />}
 
-        {people && people.length > 0 && <PeopleTable people={people}/>}
+        {people && people.length > 0 && <PeopleTable people={people} />}
       </div>
     </div>
-  )
+  );
 };
